+++ src/meico/mpm/elements/maps/ImprecisionMap.java
@@ -2,12 +2,11 @@ package meico.mpm.elements.maps;
 
 import meico.mei.Helper;
 import meico.mpm.Mpm;
-import meico.supplementary.KeyValue;
 import meico.mpm.elements.maps.data.DistributionData;
+import meico.supplementary.KeyValue;
 import meico.supplementary.RandomNumberProvider;
 import nu.xom.Attribute;
 import nu.xom.Element;
-import nu.xom.Elements;
 
 import java.util.ArrayList;
 import java.util.HashMap;