+++ src/meico/mpm/elements/styles/defs/OrnamentDef.java
@@ -5,6 +5,8 @@ import meico.mpm.Mpm;
 import nu.xom.Attribute;
 import nu.xom.Element;
 
+import java.util.ArrayList;
+
 /**
  * This class interfaces MPM's ornamentDef elements.
  * @author Axel Berndt
@@ -199,13 +201,13 @@ public class OrnamentDef extends AbstractDef {
                 if (temporalSpread.frameStart != 0.0)
                     ts.addAttribute(new Attribute("frame.start", Double.toString(temporalSpread.frameStart)));
                 if (temporalSpread.frameEnd != 0.0)
-                    ts.addAttribute(new Attribute("frame.end", Double.toString(temporalSpread.frameStart)));
+                    ts.addAttribute(new Attribute("frame.end", Double.toString(temporalSpread.frameEnd)));
                 break; 
             case Milliseconds:
                 if (temporalSpread.frameStart != 0.0)
                     ts.addAttribute(new Attribute("milliseconds.frame.start", Double.toString(temporalSpread.frameStart)));
                 if (temporalSpread.frameEnd != 0.0)
-                    ts.addAttribute(new Attribute("milliseconds.frame.end", Double.toString(temporalSpread.frameStart)));
+                    ts.addAttribute(new Attribute("milliseconds.frame.end", Double.toString(temporalSpread.frameEnd)));
                 break;
 //            case RelativeToNoteDuration:
 //                throw new UnsupportedDataTypeException("The feature TemporalSpread.FrameDomain.RelativeToNoteDuration is not yet supported.");
@@ -309,7 +311,7 @@ public class OrnamentDef extends AbstractDef {
             case "arpeg":
             case "arpeggio":
                 def.setDynamicsGradient(-1.0, 1.0);
-                def.setTemporalSpread(-180.0, 90.0, TemporalSpread.FrameDomain.Ticks, 1.0, TemporalSpread.NoteOffShift.False);
+                def.setTemporalSpread(-22.0, 22.0, TemporalSpread.FrameDomain.Ticks, 1.0, TemporalSpread.NoteOffShift.False);
         }
 
         return def;
@@ -342,6 +344,94 @@ public class OrnamentDef extends AbstractDef {
          * constructor
          */
         public TemporalSpread() {}
+
+        /**
+         * apply the temporal spread to the chord/note sequence;
+         * the notes get new attributes ornament.date.offset or ornament.date.offset.milliseconds,
+         * and ornament.duration.offset or ornament.duration.milliseconds
+         * @param chordSequence the sequence of the chords/notes in which the temporal spread is applied
+         */
+        public void apply(ArrayList<ArrayList<Element>> chordSequence) {
+            if (chordSequence.size() < 1)   // if there is no chord/note or just one
+                return;                     // we don't do anything
+
+            // process all chords/notes except for the final one
+            ArrayList<Element> previous = null;
+            if (chordSequence.size() > 1) {
+                double constFac = this.frameEnd - this.frameStart;
+                for (int i = 0; i < chordSequence.size() - 1; ++i) {    // for each chord/note until the pre-last
+                    double dateOffset = (Math.pow(((double) i) / (chordSequence.size() - 1), this.intensity) * constFac) + this.frameStart;
+                    previous = this.setOrnamentDateAtts(dateOffset, chordSequence.get(i), previous);
+                }
+            }
+
+            // place the final chord at frameEnd
+            this.setOrnamentDateAtts(this.frameEnd, chordSequence.get(chordSequence.size() - 1), previous);
+
+        }
+
+        /**
+         * helper method for method apply() to set the ornament attributes on each note:
+         *      - ornament.date.offset or ornament.milliseconds.date.offset (an offset),
+         *      - ornament.duration or ornament.milliseconds.duration (absolute duration),
+         *      - ornament.noteoff.shift (true/false)
+         * @param dateOffset
+         * @param chord
+         * @param previous the previous chord, so we can treat its duration according to the chords offset, or null
+         * @return the chord, if its duration needs treatment along the processing of the next chord (then as previous); otherwise null
+         */
+        private ArrayList<Element> setOrnamentDateAtts(double dateOffset, ArrayList<Element> chord, ArrayList<Element> previous) {
+            String dateAttName, durAttName;
+            switch (this.frameDomain) {
+                case Ticks:
+                    dateAttName = "ornament.date.offset";
+                    durAttName = "ornament.duration";
+                    break;
+                case Milliseconds:
+                    dateAttName = "ornament.milliseconds.date.offset";
+                    durAttName = "ornament.milliseconds.duration";
+                    break;
+                default:    // unknown domain
+                    return null;
+            }
+
+            // set the ornament.date.offset or ornament.milliseconds.date.offset, resp.
+            for (Element note : chord) {
+                Attribute ornamentDateAtt = Helper.getAttribute(dateAttName, note);
+                if (ornamentDateAtt != null) {
+                    dateOffset += Double.parseDouble(ornamentDateAtt.getValue());
+                    ornamentDateAtt.setValue(String.valueOf(dateOffset));
+                } else
+                    note.addAttribute(new Attribute(dateAttName, String.valueOf(dateOffset)));
+            }
+
+            // handle the ornament.duration or ornament.milliseconds.duration, resp.
+            switch (this.noteOffShift) {
+                case False:
+                    return null;
+                case True:
+                    for (Element note : chord)
+                        if (this.noteOffShift == NoteOffShift.True)
+                            note.addAttribute(new Attribute("ornament.noteoff.shift", "true"));
+                    return null;
+                case Monophonic:
+                    if (previous != null) {
+                        for (Element prev : previous) {
+                            Attribute prevDateOffsetAtt = Helper.getAttribute(dateAttName, prev);
+                            if (prevDateOffsetAtt == null)
+                                continue;
+                            Attribute ornamentDurationAtt = Helper.getAttribute(durAttName, prev);
+                            if (ornamentDurationAtt != null)
+                                ornamentDurationAtt.setValue(String.valueOf(dateOffset - Double.parseDouble(prevDateOffsetAtt.getValue())));
+                            else
+                                prev.addAttribute(new Attribute(durAttName, String.valueOf(dateOffset - Double.parseDouble(prevDateOffsetAtt.getValue()))));
+                        }
+                    }
+                    return chord;
+                default:
+                    return chord;
+            }
+        }
     }
 
     /**
@@ -356,6 +446,43 @@ public class OrnamentDef extends AbstractDef {
          * constructor
          */
         public DynamicsGradient() {}
+
+        /**
+         * apply the dynamics gradient and scale to the given chord/note sequence;
+         * the notes get a new attribute ornament.dynamics, or, if it is already present, it will be edited accordingly
+         * @param chordSequence the sequence of the chords/notes in which the dynamics gradient is applied
+         * @param scale
+         */
+        public void apply(ArrayList<ArrayList<Element>> chordSequence, double scale) {
+            if (chordSequence.size() > 1) {                                     // if we have more than one note in the chordSequence
+                double constFac = (scale * (this.transitionTo - this.transitionFrom)) / (chordSequence.size() - 1);
+                double fromVelocity = this.transitionFrom * scale;
+                for (int n = 0; n < chordSequence.size(); ++n) {                // for each chord in the list
+                    double ornamentDynamics = (constFac * n) + fromVelocity;    // compute its velocity (the value is relative to the basic dynamics)
+                    this.setOrnamentDynamicsAtt(ornamentDynamics, chordSequence.get(n));
+                }
+            } else if (chordSequence.size() > 0) {                              // if there is only one chord/note in the chordSequence
+                double ornamentDynamics = this.transitionTo * scale;
+                this.setOrnamentDynamicsAtt(ornamentDynamics, chordSequence.get(0));
+            }
+        }
+
+        /**
+         * helper method for method apply() to set the ornament.dynamics attribute on each note in the given chord
+         * @param ornamentDynamics
+         * @param chord
+         */
+        private void setOrnamentDynamicsAtt(double ornamentDynamics, ArrayList<Element> chord) {
+            for (Element note : chord) {
+                Attribute ornamentDynamicsAtt = Helper.getAttribute("ornament.dynamics", note);
+                if (ornamentDynamicsAtt != null) {                                          // if there is already an ornament.dynamics attribute
+                    ornamentDynamics += Double.parseDouble(ornamentDynamicsAtt.getValue()); // add the values
+                    ornamentDynamicsAtt.setValue(String.valueOf(ornamentDynamics));
+                } else {
+                    note.addAttribute(new Attribute("ornament.dynamics", String.valueOf(ornamentDynamics)));
+                }
+            }
+        }
     }
 
 }